pyproject.toml
[project]
name = "ai-notes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "uvicorn>=0.44.0",
    "whisperx>=3.8.5",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]
